requirements.txt
asgiref==3.4.1
black==21.8b0
click==8.0.1
colorama==0.4.4
Django==3.2.7
djangorestframework==3.12.4
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
pytz==2021.1
qrcode==7.3
regex==2021.8.28
sqlparse==0.4.1
tomli==1.2.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
